+++ CHANGES.md
@@ -3,6 +3,8 @@ Change Log
 
 ## changes in 1.3-SNAPSHOT
 
+* added clampToGround to GeoJsonDataSource 
+
 
 ## changes in 1.2
 
@@ -54,5 +56,5 @@ to extends ImageryProvider, also add a corresponding object to all that extends 
 * added aircraftHeadingPitchRollToFixedFrame, aircraftHeadingPitchRollQuaternion to Transforms
 * renamed headingPitchRollMatrix4 to headingPitchRollToFixedFrame in Transforms
 * added heightReference to PointGraphics and Model (Billboard already has it)
-* added clampToGround to KMLDataSource and GeoJsonDataSource in CesiumOptions
+* added clampToGround to KMLDataSourceOptions and GeoJsonDataSourceOptions 
 * added unsupportedNodeEvent to KMLDataSource in Cesium